Lisasin Transactionile LoanAccounti.

+++ Pages/Client/TransactionPage.cs
@@ -32,14 +32,14 @@ namespace BankingApp.Pages.Client
 
         public override string GetName(IHtmlHelper<TransactionPage> h, int i) => i switch
         {
-            6 => base.getName<int>(h, i),
+            6 => base.getName<double>(h, i),
             7 => base.getName<int>(h, i),
             _ => base.GetName(h, i)
         };
 
         public override IHtmlContent GetValue(IHtmlHelper<TransactionPage> h, int i) => i switch
         {
-            6 => base.getValue<int>(h, i),
+            6 => base.getValue<double>(h, i),
             7 => base.getValue<int>(h, i),
             _ => base.GetValue(h, i)
         };
